gnatbus: fix shutdown request
Calling pause_all_vcpus() is switching from the vcpu-thread to the cpu-thread.
So if we have a shutdown packet in a read or write handler this is likely to
make the reply packet to be received by the io-thread instead of the
send_and_wait_resp() function in the vcpu-thread. As a result the packet is
not recognized and we deadlock in send_and_wait_resp(). So instead of calling
pause_all_vcpus() directly just wait the reply packet to be received and call
it later. If we get the packet asynchronously we call it directly.

TN: R907-022
Change-Id: I0e6f71e4009d004dc189354436321f277182f154