ppc: mpc8641d: use a bottom half to reset the board
The following guest code might deadlock under heavy loaded machine:

  _exit:
    write to a MMIO reset
  loop:
    branch loop

So resetting / stopping the CPU from the vcpu thread might be delicate
because the rest of the translation code is executed. So lets create a
bottom half in order to reset the board so the IO thread will notify
the VCPU thread and reset the board.

TN: RA25-011
Change-Id: Iddb292a502dcc20bac15cc993f01c27628044305